refactor(cliproxy): extract harness templates into setup/templates.ts
Move harnessSchema, Harness type, SecretAssignment, VariableAssignment,
HarnessTemplate, getHarnessTemplate, formatTemplateSummary, collectCollisions,
and stripTrailingSlash into a focused submodule. Lifts OMO_TOKEN out of the
getHarnessTemplate body to a module-private const. Byte-identical anthropic-only
output preserved (regression tests moved with the symbols).

setup.ts shrinks from 1408 to 1261 LOC. Adds 6 new TDD-first unit tests for
stripTrailingSlash, collectCollisions, and formatTemplateSummary (698 total, +14).
