+++ packages/cli/src/commands/cliproxy/setup.test.ts
@@ -8,7 +8,6 @@ import {
   buildNonInteractivePlan,
   formatDryRunPreview,
   formatWorkflowSnippet,
-  getHarnessTemplate,
   interpretGhContentResult,
   isGhRateLimitError,
   mustConfirmDestructive,
@@ -17,9 +16,8 @@ import {
   validateSetupOptions,
   verifyModelsAvailable,
   withGhRetry,
-  type SecretAssignment,
-  type VariableAssignment,
 } from './setup'
+import {getHarnessTemplate} from './setup/templates'
 
 const COMPLETE_WORKFLOW = `      - uses: fro-bot/agent@abc123
         with:
@@ -156,49 +154,6 @@ describe('cliproxy setup helpers', () => {
       expect(() => validateSetupOptions({key: 'sk-test', repo: 'owner/repo'}, false)).toThrow(
         '--harness is required when stdin is not a TTY',
       )
-    })
-  })
-
-  describe('getHarnessTemplate', () => {
-    it('returns the expected OpenCode secret and variable names', () => {
-      const template = getHarnessTemplate('opencode')
-
-      expect(template.secrets.map((entry: SecretAssignment) => entry.name)).toEqual([
-        'OPENCODE_AUTH_JSON',
-        'OPENCODE_CONFIG',
-        'OMO_PROVIDERS',
-      ])
-      expect(template.variables.map((entry: VariableAssignment) => entry.name)).toEqual(['FRO_BOT_MODEL'])
-    })
-
-    it('uses a provider-prefixed FRO_BOT_MODEL default value', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'sk-test'})
-      const modelEntry = template.variables.find((entry: VariableAssignment) => entry.name === 'FRO_BOT_MODEL')
-
-      expect(modelEntry?.value).toMatch(/^anthropic\//)
-    })
-
-    it('uses the expected OMO_PROVIDERS default value', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'sk-test'})
-      const providersEntry = template.secrets.find((entry: SecretAssignment) => entry.name === 'OMO_PROVIDERS')
-
-      expect(providersEntry?.value).toBe('claude-max20')
-    })
-
-    it('writes an OPENCODE_CONFIG baseURL with the /v1 suffix', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'sk-test'})
-      const configEntry = template.secrets.find((entry: SecretAssignment) => entry.name === 'OPENCODE_CONFIG')
-      const parsed = JSON.parse(configEntry?.value ?? '{}')
-
-      expect(parsed.provider.anthropic.options.baseURL).toMatch(/\/v1$/)
-    })
-
-    it('writes OPENCODE_AUTH_JSON with type=api and the supplied key', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'sk-test-key'})
-      const authEntry = template.secrets.find((entry: SecretAssignment) => entry.name === 'OPENCODE_AUTH_JSON')
-      const parsed = JSON.parse(authEntry?.value ?? '{}')
-
-      expect(parsed.anthropic).toEqual({type: 'api', key: 'sk-test-key'})
     })
   })
 
@@ -501,202 +456,6 @@ describe('option parsing', () => {
 
     it('rejects "openai/" (empty tail)', () => {
       expect(MODEL_RE.test('openai/')).toBe(false)
-    })
-  })
-})
-
-describe('getHarnessTemplate provider-aware', () => {
-  // Frozen byte-identical string for the anthropic-only regression test.
-  // This is the EXACT output of getHarnessTemplate('opencode', {keyValue: 'test-key'})
-  // baseline anthropic-only output. Any change to this string is a breaking regression.
-  const ANTHROPIC_ONLY_AUTH_JSON = '{"anthropic":{"type":"api","key":"test-key"}}'
-  const ANTHROPIC_ONLY_CONFIG = '{"provider":{"anthropic":{"options":{"baseURL":"https://cliproxy.fro.bot/v1"}}}}'
-
-  describe('regression — anthropic-only (byte-identical)', () => {
-    it('no providers/model args → OPENCODE_AUTH_JSON is byte-identical to baseline', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'test-key'})
-      const authEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-
-      expect(authEntry?.value).toBe(ANTHROPIC_ONLY_AUTH_JSON)
-    })
-
-    it('no providers/model args → OPENCODE_CONFIG is byte-identical to baseline', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'test-key'})
-      const configEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_CONFIG')
-
-      expect(configEntry?.value).toBe(ANTHROPIC_ONLY_CONFIG)
-    })
-
-    it('no providers/model args → OMO_PROVIDERS is claude-max20', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'test-key'})
-      const entry = template.secrets.find((e: SecretAssignment) => e.name === 'OMO_PROVIDERS')
-
-      expect(entry?.value).toBe('claude-max20')
-    })
-
-    it('no providers/model args → FRO_BOT_MODEL is anthropic/claude-sonnet-4-6', () => {
-      const template = getHarnessTemplate('opencode', {keyValue: 'test-key'})
-      const entry = template.variables.find((e: VariableAssignment) => e.name === 'FRO_BOT_MODEL')
-
-      expect(entry?.value).toBe('anthropic/claude-sonnet-4-6')
-    })
-
-    it("explicit providers: ['anthropic'] → byte-identical to no-providers output", () => {
-      const baseline = getHarnessTemplate('opencode', {keyValue: 'test-key'})
-      const explicit = getHarnessTemplate('opencode', {keyValue: 'test-key', providers: ['anthropic']})
-
-      const baselineAuth = baseline.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-      const explicitAuth = explicit.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-      expect(explicitAuth?.value).toBe(baselineAuth?.value)
-
-      const baselineConfig = baseline.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_CONFIG')
-      const explicitConfig = explicit.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_CONFIG')
-      expect(explicitConfig?.value).toBe(baselineConfig?.value)
-    })
-  })
-
-  describe('openai-only provider', () => {
-    it("providers: ['openai'], model: 'openai/gpt-5.4-mini' → correct OPENCODE_AUTH_JSON", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-openai-key',
-        providers: ['openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const authEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-
-      expect(authEntry?.value).toBe('{"openai":{"type":"api","key":"sk-openai-key"}}')
-    })
-
-    it("providers: ['openai'], model: 'openai/gpt-5.4-mini' → correct OPENCODE_CONFIG", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-openai-key',
-        providers: ['openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const configEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_CONFIG')
-
-      expect(configEntry?.value).toBe('{"provider":{"openai":{"options":{"baseURL":"https://cliproxy.fro.bot/v1"}}}}')
-    })
-
-    it("providers: ['openai'], model: 'openai/gpt-5.4-mini' → OMO_PROVIDERS is openai", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-openai-key',
-        providers: ['openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const entry = template.secrets.find((e: SecretAssignment) => e.name === 'OMO_PROVIDERS')
-
-      expect(entry?.value).toBe('openai')
-    })
-
-    it("providers: ['openai'], model: 'openai/gpt-5.4-mini' → FRO_BOT_MODEL is openai/gpt-5.4-mini", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-openai-key',
-        providers: ['openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const entry = template.variables.find((e: VariableAssignment) => e.name === 'FRO_BOT_MODEL')
-
-      expect(entry?.value).toBe('openai/gpt-5.4-mini')
-    })
-
-    it("providers: ['openai'] with no model → uses PROVIDER_DEFAULTS openai/gpt-5.4-mini", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-openai-key',
-        providers: ['openai'],
-      })
-      const entry = template.variables.find((e: VariableAssignment) => e.name === 'FRO_BOT_MODEL')
-
-      expect(entry?.value).toBe('openai/gpt-5.4-mini')
-    })
-  })
-
-  describe('dual-provider (anthropic + openai)', () => {
-    it("providers: ['anthropic', 'openai'] → OPENCODE_AUTH_JSON has anthropic-first key order", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-dual',
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const authEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-
-      expect(authEntry?.value).toBe(
-        '{"anthropic":{"type":"api","key":"sk-dual"},"openai":{"type":"api","key":"sk-dual"}}',
-      )
-    })
-
-    it("providers: ['anthropic', 'openai'] → OPENCODE_CONFIG has anthropic-first key order", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-dual',
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const configEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_CONFIG')
-
-      expect(configEntry?.value).toBe(
-        '{"provider":{"anthropic":{"options":{"baseURL":"https://cliproxy.fro.bot/v1"}},"openai":{"options":{"baseURL":"https://cliproxy.fro.bot/v1"}}}}',
-      )
-    })
-
-    it("providers: ['anthropic', 'openai'] → OMO_PROVIDERS is claude-max20,openai", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-dual',
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const entry = template.secrets.find((e: SecretAssignment) => e.name === 'OMO_PROVIDERS')
-
-      expect(entry?.value).toBe('claude-max20,openai')
-    })
-
-    it("providers: ['anthropic', 'openai'] → FRO_BOT_MODEL is the supplied model", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-dual',
-        providers: ['anthropic', 'openai'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const entry = template.variables.find((e: VariableAssignment) => e.name === 'FRO_BOT_MODEL')
-
-      expect(entry?.value).toBe('openai/gpt-5.4-mini')
-    })
-
-    it("providers: ['openai', 'anthropic'] (openai first) → output is still anthropic-first in JSON", () => {
-      const template = getHarnessTemplate('opencode', {
-        keyValue: 'sk-dual',
-        providers: ['openai', 'anthropic'],
-        model: 'openai/gpt-5.4-mini',
-      })
-      const authEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-
-      expect(authEntry?.value).toBe(
-        '{"anthropic":{"type":"api","key":"sk-dual"},"openai":{"type":"api","key":"sk-dual"}}',
-      )
-    })
-
-    it('multiple providers with no model → throws "model required when multiple providers selected"', () => {
-      expect(() =>
-        getHarnessTemplate('opencode', {
-          keyValue: 'sk-dual',
-          providers: ['anthropic', 'openai'],
-        }),
-      ).toThrow('model required when multiple providers selected')
-    })
-  })
-
-  describe('edge cases', () => {
-    it('keyValue: undefined → auth-json key is sk-placeholder', () => {
-      const template = getHarnessTemplate('opencode', {providers: ['anthropic']})
-      const authEntry = template.secrets.find((e: SecretAssignment) => e.name === 'OPENCODE_AUTH_JSON')
-      const parsed = JSON.parse(authEntry?.value ?? '{}')
-
-      expect(parsed.anthropic.key).toBe('sk-placeholder')
-    })
-
-    it('claude-code harness is unaffected by providers/model args', () => {
-      const template = getHarnessTemplate('claude-code', {keyValue: 'sk-cc'})
-
-      expect(template.secrets).toHaveLength(1)
-      expect(template.secrets[0]?.name).toBe('ANTHROPIC_API_KEY')
     })
   })
 })
